api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-jobspy>=1.1.0
pypdf>=3.17.0
google-generativeai>=0.3.0
requests>=2.31.0
duckduckgo-search>=5.0.0
